Alters account detail section to always be a form, adds styling
